# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_basilex_promenade_pkg_response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_basilex_promenade_pkg_response.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_basilex_promenade_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Error'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Pagination'
      status:
        type: string
    type: object
...
  internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      currency_code:
        type: string
      from_location_code:
        type: string
      from_warehouse_id:
        type: string
      id:
        type: string
      inventory_id:
        type: string
      movement_date:
        type: string
      notes:
        type: string
      quantity:
        type: integer
      quantity_after_move:
        type: integer
      quantity_before_move:
        type: integer
      reason:
        type: string
      reference_id:
        type: string
      reference_type:
        type: string
      to_location_code:
        type: string
      to_warehouse_id:
        type: string
      total_cost_cents:
        type: integer
      type:
        type: string
      unit_cost_cents:
        type: integer
    type: object
  internal_contexts_warehouse_stockmovement_adapter_http.RecordMovementRequest:
    properties:
      created_by:
        type: string
      inventory_id:
        type: string
      quantity:
        type: integer
      quantity_before:
        description: Current quantity before movement
        type: integer
      type:
        type: string
    required:
    - created_by
    - inventory_id
    - quantity
    - quantity_before
    - type
    type: object
...
host: localhost:8081
info:
  contact:
    email: alexander.vasilenko@gmail.com
    name: API Support
    url: https://github.com/basilex/promenade
  description: Modern backend platform for customer management, orders, and business
    workflows with clean DDD architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Promenade Platform
  version: 0.1.0
paths:
...
  /warehouse/stock-movements:
    post:
      consumes:
      - application/json
      description: Records a generic stock movement (receipt, reservation, commit,
        adjustment, transfer, damage, return)
      parameters:
      - description: Stock movement details
        in: body
        name: movement
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.RecordMovementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Movement recorded successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Record stock movement
      tags:
      - Stock Movements
  /warehouse/stock-movements/{id}:
    get:
      description: Retrieves stock movement details by its unique identifier
      parameters:
      - description: Movement ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Movement details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
              type: object
        "400":
          description: Invalid movement ID
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Movement not found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get stock movement by ID
      tags:
      - Stock Movements
  /warehouse/stock-movements/inventory/{inventory_id}:
    get:
      description: Retrieves paginated stock movements for a specific inventory item
      parameters:
      - description: Inventory ID
        format: uuid
        in: path
        name: inventory_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated movements
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  properties:
                    movements:
                      items:
                        $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
                      type: array
                    page:
                      type: integer
                    page_size:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Invalid inventory ID
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get stock movements by inventory
      tags:
      - Stock Movements
  /warehouse/stock-movements/recent:
    get:
      description: Retrieves the 50 most recent stock movements across all inventory
        items
      produces:
      - application/json
      responses:
        "200":
          description: Recent movements
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get recent stock movements
      tags:
      - Stock Movements
  /warehouse/stock-movements/reference:
    get:
      description: Retrieves stock movements linked to a specific reference (order,
        PO, adjustment)
      parameters:
      - description: Reference type (order, po, adjustment)
        in: query
        name: type
        required: true
        type: string
      - description: Reference ID
        format: uuid
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Movements for reference
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
                  type: array
              type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get stock movements by reference
      tags:
      - Stock Movements
  /warehouse/stock-movements/summary/{inventory_id}:
    get:
      description: Retrieves total in, total out, and net change for an inventory
        (last 30 days by default)
      parameters:
      - description: Inventory ID
        format: uuid
        in: path
        name: inventory_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Movement summary
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  properties:
                    end_date:
                      type: string
                    net_change:
                      type: integer
                    start_date:
                      type: string
                    total_in:
                      type: integer
                    total_out:
                      type: integer
                  type: object
              type: object
        "400":
          description: Invalid inventory ID
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get inventory movement summary
      tags:
      - Stock Movements
  /warehouse/stock-movements/type/{type}:
    get:
      description: Retrieves paginated stock movements of a specific type (last 30
        days by default)
      parameters:
      - description: Movement type (receipt, reservation, reservation_release, commit,
          adjustment, transfer, damage, return)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated movements
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  properties:
                    movements:
                      items:
                        $ref: '#/definitions/internal_contexts_warehouse_stockmovement_adapter_http.MovementResponse'
                      type: array
                    page:
                      type: integer
                    page_size:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get stock movements by type
      tags:
      - Stock Movements
schemes:
- http
- https
swagger: "2.0"